Cleanup cypress test file names and organization
