+++ front-end/cypress/e2e/F3X/disbursements.cy.ts
@@ -9,12 +9,14 @@ import {
 import {
   defaultScheduleFormData as defaultTransactionFormData,
   DisbursementFormData,
+  formTransactionDataForSchedule,
 } from '../models/TransactionFormModel';
 import { F3XSetup } from './f3x-setup';
 import { StartTransaction } from './start-transaction/start-transaction';
 import { faker } from '@faker-js/faker';
 import { ReportListPage } from '../pages/reportListPage';
 import { F24Setup } from '../F24/f24-setup';
+import { ContactListPage } from '../pages/contactListPage';
 
 const independentExpVoidData: DisbursementFormData = {
   ...defaultTransactionFormData,
@@ -98,5 +100,82 @@ describe('Disbursements', () => {
     PageUtils.clickLink('Independent Expenditure');
     cy.contains(individualContactFormData.first_name).should('exist');
     cy.contains(individualContactFormData.last_name).should('exist');
+  });
+
+  it('Create an Other Disbursement transaction', () => {
+    F3XSetup();
+    StartTransaction.Disbursements().Other().Other();
+
+    PageUtils.clickLink('Create a new contact');
+    const formContactData = {
+      ...individualContactFormData,
+      ...{ contact_type: 'Organization' },
+    };
+    ContactListPage.enterFormData(formContactData, true);
+    PageUtils.clickButton('Save & continue');
+
+    const formTransactionData = {
+      ...defaultTransactionFormData,
+      ...{
+        amount: 200.01,
+        category_code: '005 Polling Expenses',
+        electionYear: undefined,
+        electionType: undefined,
+        date_received: new Date(currentYear, 4 - 1, 27),
+      },
+    };
+    TransactionDetailPage.enterScheduleFormData(formTransactionData);
+    PageUtils.clickButton('Save');
+    cy.contains('Confirm').should('exist');
+    PageUtils.clickButton('Continue');
+
+    cy.get('tr').should('contain', 'Other Disbursement');
+    cy.get('tr').should('not.contain', 'Unitemized');
+    cy.get('tr').should('contain', formContactData.name);
+    cy.get('tr').should('contain', PageUtils.dateToString(formTransactionData.date_received));
+    cy.get('tr').should('contain', '$' + formTransactionDataForSchedule.amount);
+
+    // Check values of edit form
+    PageUtils.clickLink('Other Disbursement');
+    cy.get('#entity_type_dropdown > div.readonly').should('exist');
+    cy.get('#entity_type_dropdown').should('contain', 'Organization');
+    ContactListPage.assertFormData(formContactData, true);
+    TransactionDetailPage.assertFormData(formTransactionDataForSchedule);
+  });
+
+  it('Create a Credit Card Payment for 100% Federal Election Activity transaction', () => {
+    F3XSetup({ organization: true });
+    setCommitteeType('X');
+    StartTransaction.Disbursements().Federal().CreditCardPayment();
+
+    cy.get('[id="searchBox"]').type(organizationFormData.name.slice(0, 1));
+    cy.contains(organizationFormData.name).should('exist');
+    cy.contains(organizationFormData.name).click();
+
+    const transactionFormData = {
+      ...formTransactionDataForSchedule,
+      ...{
+        electionType: 'General',
+        electionYear: 2024,
+        election_other_description: faker.lorem.sentence({ min: 1, max: 2 }),
+        purpose_description: '',
+        category_code: '',
+        date_received: new Date(currentYear, 4 - 1, 27),
+      },
+    };
+    TransactionDetailPage.enterScheduleFormData(transactionFormData, false, '', false);
+    cy.get('[data-test="navigation-control-button"]').contains('button', 'Save').click();
+
+    cy.get('tr').should('contain', 'Credit Card Payment for 100% Federal Election Activity');
+    cy.get('tr').should('contain', organizationFormData['name']);
+    cy.get('tr').should('contain', PageUtils.dateToString(transactionFormData.date_received));
+    cy.get('tr').should('contain', '$' + transactionFormData.amount);
+
+    // Check values of edit form
+    PageUtils.clickLink('Credit Card Payment for 100% Federal Election Activity');
+    cy.get('#entity_type_dropdown > div.readonly').should('exist');
+    cy.get('#entity_type_dropdown').should('contain', 'Organization');
+    ContactListPage.assertFormData(organizationFormData, true);
+    TransactionDetailPage.assertFormData(transactionFormData);
   });
 });
